recast current_rating to float

+++ src/foosball/backend/database/migrations/20191101110301_users.js
@@ -3,7 +3,7 @@ exports.up = function (knex, Promise) {
       table.increments()
       table.string('first_name').notNullable().defaultsTo('')
       table.string('last_name').notNullable().defaultsTo('')
-      table.integer('current_rating').notNullable().defaultsTo(1000)
+      table.float('current_rating').notNullable().defaultsTo(1000)
       table.integer('games_played').notNullable().defaultsTo(0)
       table.timestamps(true, true)
   })
